Add user articles view
Allows the user to view their own posted articles, and as of now, delete
them. The list all articles view has been modified to not permit the
deletion of articles; this can only be achieved from the new user view.
Minor modifications to fit the new server API response, necessary to
implement this new feature