Add support for multiple article pictures
Extensively modifies the Edit Article activity and adds new activities
and layouts to support an article having more than one picture.

The logic for saving the pictures is modified: the backend will now
store the direct image URLs, and they will be served directly instead of
reading from the Firebase storage.

The UI for multiple images is implemented with a Slider library. The
selection is handled from another activity with a Grid View, that uses
an Image adapter to dynamically add and remove images, and the final
state is passed on as an array of URLs back to the main edit article
activity.

The Article model is now always initialzed with null values as soon as
the edit article activity is created, and values are read from the UI
and passed on to setters when the user clicks save. This allows us to
store pictures on the model directly, whenever the user decides to add
some, instead of keeping a temporary array and passing it onto the model
when we need to hit the backend. Some Text Watcher logic to
enable/disable the save button was modified to allow this change.

With this change, the FirebaseImageManager util class becomes a little
obsolete, since we store the image URLs directly, we don't need to go
through the storage. Maybe the update account logic should follow the
same pattern of storing URLs to fully deprecate the module.