QR codes generation and reading
Adds an option to the Article Edit activity to allow the owner to
generate a QR core for their article. Adds a side panel option to scan
bar codes. When a user scans the article bar code, they'll be redirected
to the article detail activity of the corresponding article.

Modifies some logic in BaseActivity to better generalize actions to
perform after clicking an option in the side panel.